refactor: implementing infra layer in modules
