+++ src/modules/customers/services/ShowCustomerService.ts
@@ -1,6 +1,6 @@
 import { getCustomRepository } from "typeorm"
-import CustomersRepository from "../typeorm/repositories/CustomersRepository";
-import Customer from "@modules/customers/typeorm/entities/Customer";
+import CustomersRepository from "../infra/typeorm/repositories/CustomersRepository";
+import Customer from "@modules/customers/infra/typeorm/entities/Customer";
 import AppError from "@shared/errors/AppError";
 
 interface IRequest {
