+++ src/modules/orders/infra/http/controllers/OrdersController.ts
@@ -1,6 +1,6 @@
 import { Request, Response } from "express";
-import CreateOrderService from "../services/CreateOrderService";
-import ShowOrderService from "../services/ShowOrderService";
+import CreateOrderService from "../../../services/CreateOrderService";
+import ShowOrderService from "../../../services/ShowOrderService";
 
 export default class OrdersController {
