+++ src/modules/orders/infra/typeorm/entities/OrderProduct.ts
@@ -1,4 +1,4 @@
-import Product from '@modules/products/typeorm/entities/Product';
+import Product from '@modules/products/infra/typeorm/entities/Product';
 import { Entity, PrimaryGeneratedColumn, Column, CreateDateColumn, UpdateDateColumn, ManyToMany, ManyToOne, JoinColumn } from 'typeorm'
 import Order from './Order';
