+++ src/modules/orders/infra/typeorm/repositories/OrdersRepository.ts
@@ -1,4 +1,4 @@
-import Customer from "@modules/customers/typeorm/entities/Customer";
+import Customer from "@modules/customers/infra/typeorm/entities/Customer";
 import { EntityRepository, Repository } from "typeorm";
 import Order from "../entities/Order";
