+++ src/modules/products/infra/http/controllers/ProductsController.ts
@@ -1,9 +1,9 @@
 import { Request, Response } from "express";
-import CreateProductService from "../services/CreateProductService";
-import ListProductService from "../services/ListProductService";
-import ShowProductService from "../services/ShowProductService";
-import UpdateProductService from "../services/UpdateProductService";
-import DeleteProductService from "../services/DeleteProductService";
+import CreateProductService from "../../../services/CreateProductService";
+import ListProductService from "../../../services/ListProductService";
+import ShowProductService from "../../../services/ShowProductService";
+import UpdateProductService from "../../../services/UpdateProductService";
+import DeleteProductService from "../../../services/DeleteProductService";
 
 export default class ProductsController {
