+++ src/modules/products/services/ShowProductService.ts
@@ -1,6 +1,6 @@
 import { getCustomRepository } from "typeorm"
-import { ProductRepository } from "../typeorm/repositories/ProductsRepository";
-import Product from "@modules/products/typeorm/entities/Product";
+import { ProductRepository } from "../infra/typeorm/repositories/ProductsRepository";
+import Product from "@modules/products/infra/typeorm/entities/Product";
 import AppError from "@shared/errors/AppError";
 
 interface IRequest{
