+++ src/modules/users/infra/http/controllers/ForgotPasswordController.ts
@@ -1,5 +1,5 @@
 import { Request, Response } from 'express'
-import SendForgotPasswordEmailService from '../services/SendForgotPasswordEmailService';
+import SendForgotPasswordEmailService from '../../../services/SendForgotPasswordEmailService';
 
 export default class ForgotPasswordController {
