+++ src/shared/infra/http/routes/index.ts
@@ -1,10 +1,10 @@
 import { Router } from 'express';
-import productsRouter from '@modules/products/routes/products.routes';
-import usersRouter from '@modules/users/routes/users.routes';
-import customersRouter from '@modules/customers/routes/customers.routes';
-import ordersRouter from '@modules/orders/routes/order.routes';
-import authRouter from '@modules/users/routes/auth.routes';
-import passwordRouter from '@modules/users/routes/password.routes';
+import productsRouter from '@modules/products/infra/http/routes/products.routes';
+import usersRouter from '@modules/users/infra/http/routes/users.routes';
+import customersRouter from '@modules/customers/infra/http/routes/customers.routes';
+import ordersRouter from '@modules/orders/infra/http/routes/order.routes';
+import authRouter from '@modules/users/infra/http/routes/auth.routes';
+import passwordRouter from '@modules/users/infra/http/routes/password.routes';
 
 const routes = Router();
